+++ src/utils.ts
@@ -332,12 +332,5 @@ export const getColumnTestConfigFromYml = (
 };
 
 export function getFormattedDateTime(): string {
-  const now = new Date();
-
-  const date = now.toLocaleDateString("en-GB").replace(/\//g, "-");
-  const time = now
-    .toLocaleTimeString("en-GB", { hour12: false })
-    .replace(/:/g, "-");
-
-  return `${date}__${time}`;
+  return new Date().toLocaleString().replace(/[/ ]/g, "-");
 }
