+++ webview_panels/src/modules/queryPanel/components/runAdhocQueryButton/RunAdhocQueryButton.tsx
@@ -7,7 +7,7 @@ const RunAdhocQueryButton = (): JSX.Element => {
   };
   return (
     <Button outline onClick={handleClick}>
-      Run adhoc query
+      + New query
     </Button>
   );
 };
